+++ src/app/content.rs
@@ -89,15 +89,6 @@ impl InstructionListStates {
         self.last_index = current_instruction_idx - 1;
     }
 
-    /// Used to set the instruction that should be executed next.
-    ///
-    /// Updates the visuals, current and last index.
-    pub fn set_instruction(&mut self, index: usize) {
-        self.force_set(index);
-        self.last_index = i32::try_from(index).expect("Unable to parse i32 from usize.");
-        self.current_index = i32::try_from(index + 1).expect("Unable to parse i32 from usize.");
-    }
-
     /// Used to force the highlight of a specific line.
     ///
     /// Should only be used for visuals if the line is known that should be highlighted.